mover la funcion de feed
